split up model spec
* add coverage spec
* remove duplicate attribute tests
