+++ spec/attributes_spec.rb
@@ -163,14 +163,4 @@ describe Cistern::Attributes, 'parsing' do
   it 'should slice out unaccounted for attributes' do
     expect(TypeSpec.new({ 'something' => { 'id' => '12' } }).attributes.keys).not_to include('something')
   end
-
-  describe '#requires' do
-    it 'should raise if attribute not provided' do
-      expect { TypeSpec.new({ 'service' => 'fake', 'something' => { 'id' => '12' } }).save }.to raise_exception(ArgumentError)
-    end
-
-    it 'should raise if attribute is provided and is nil' do
-      expect { TypeSpec.new({ 'service' => 'fake', 'custom' => nil }).save }.to raise_exception(ArgumentError)
-    end
-  end
 end
